pyproject.toml
[project]
requires-python = ">=3.12"
name = "real_estate_analisis"

[tool.poetry]
version = "0.1.0"
description = ""
readme = "README.md"
authors = []

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi=">=0.115.11,<0.116.0"
pydantic-settings=">=2.8.1,<3.0.0"
uvicorn=">=0.34.0,<0.35.0"
prometheus-client = "^0.21.1"
pyyaml = "^6.0.2"
aiohttp = "^3.12.13"
aiofiles = "^24.1.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.43"
fastapi-babel = "^1.0.0"
python-multipart = "^0.0.20"
playwright = "^1.55.0"
psutil = "^7.1.0"
bs4 = "^0.0.2"
regex = "^2025.9.18"
apscheduler = "^3.11.0"
billiard = "^4.2.2"
alembic = "^1.17.2"

[tool.poetry.group.dev.dependencies]
pytest=">=8.3.5,<9.0.0"
pytest-asyncio=">=0.25.3,<0.26.0"
requests=">=2.32.3,<3.0.0"
httpx=">=0.28.1,<0.29.0"
pytest-cov=">=6.0.0,<7.0.0"
black = "^25.1.0"

[tool.pytest.ini_options]
pythonpath = ". app"
asyncio_mode = "auto"
addopts = "--disable-warnings"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

